pyproject.toml
[project]
name = "hiring-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.100.2",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "typing-extensions>=4.14.1",
]
